feat: refactor lint-staged configurations and add shared utilities for workspace support
